auto-bot-producer/requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
pydantic==2.11.4
nest-asyncio==1.6.0 